I tested it on WINDOWS! Had to fix some compiling errors in which I got no idea how gcc accepted them... I'm on my laptop, there're still problems with rounding on d3d11 there... :(

+++ src/core/drivers/direct3d11/d3d11draw.cpp
@@ -108,7 +108,7 @@ void Instance::flush_queueinstances2(Texture* texture, const sre::RenderInstance
         m_dxdevicecontext->PSSetShaderResources(0, 1, texture ? &texture->dxsrv : &m_basictexture);
     }
 
-    m_dxdevicecontext->Draw(point_count, 0);
+    m_dxdevicecontext->Draw(static_cast<UINT>(point_count), 0);
 }
 
 //
